requirements.txt
streamlit
faiss-cpu
sentence-transformers
langchain
unstructured
groq
langchain-community
